# Conflicts:
#	com.unity.render-pipelines.universal/Editor/UniversalRenderPipelineCameraEditor.cs

+++ Tools/git-hook/preinstall.py
@@ -1,10 +1,12 @@
 #!/usr/bin/python
 
 import sys, os, subprocess, shutil
+from os import path as os_path
 
 
 def clean_node_modules():
-	shutil.rmtree("./node_modules")
+	if os_path.exists("./node_modules"):
+	  shutil.rmtree("./node_modules")
 
 
 # Sometimes git hooks are already existing in .git/hooks. 
@@ -18,6 +20,9 @@ def save_existing_hooks():
 		pre_installed_hook = path + hook
 		tmp_hook = "./git-hook/tmp_" + hook
 
+		if not os_path.exists(pre_installed_hook):
+			continue
+
 		with open(pre_installed_hook, 'r') as pre_installed_hook_file:
   			hook_data = pre_installed_hook_file.read()	  
 
@@ -28,5 +33,6 @@ def save_existing_hooks():
 
 
 if __name__== "__main__":
+	print(f'Any question about the hooks or problem with the installation? Take a look at the FAQ in Tools/readme.md.\n')
 	clean_node_modules()
 	save_existing_hooks()
